[#14] AUV inspecting a monopile
MILESTONE: the AUV can follow an hexagonal path.
TODOs: the path is following "looking ahead", it needs to be transformed as looking inside instead.
Line 60 needs to be updated with +-pi/2.
Next, a variable velocity needs to be imposed, i.e. u=0, v=0.5 after reaching the monopile.
Next, include full faults on the thrusters.
Next include the ZOH.