[#14] AUV inspecting a monopile
Introduced a two stage guidance:
a) approaching the orbit;
b) on the orbit.

In phase a), the AUV moves with a constant surge speed, and no sway. In pahse b), the AUV point towards the monopile and moves
with a constant sway velocity.

Introduced further parameters and automated the generation of the waypoints.