[#28] Gravity model with point gravity and with constant gravity acceleration
Created custom world model.
The g=9.81 m/s2 (enforced). The SML proposes a model where g is proportional to 1/r^2 (with r the distance from the centre of the world). This is a nice idea, but the model keep causing instabilities.
I am convinced it is due to the high values of the r and mu (another gravity related parameter), which span e06 and e14.

I enforced g=9.81 everywhere now, and it works.