[#29] Milestones
1) The model has all the frames fixed (CAVEAT: in this commit, the NED frame is oriented werdly, becaused I forced
a "wrong position": nothing major, just NED position and orientation to be fixed).

2) The buoyancy acts in the global frame, compensating for the gravity exactly (there are some numerical errors:
I am convinced they are linked to the initial conditions, which force the body to have a tiny, but non-zero
acceleration, e.g. mm/s2. This acceleration compounds over the course of 1 day in simulation time, and generate a
diplacement of {~0m, ~32m, ~-4mm}.

TODO next:
a) Fix the automatic position and orientation of the NED frame, which is now provided manually. All the equations are already into
the Kinematics.ReferenceFrames.

b) Update the buoyancy model to include the volume and not the artificial mass, as it right now.